requirements_for_rpi.txt
keras==2.10.0
librosa==0.8.0
numpy==1.21.5
pandas==1.1.3
PyYAML==6.0.1
soundfile==0.12.1
tensorflow==2.10.0
tqdm==4.54.1
